Partition ATSWriter to write xml fragments in separate source files
* Add shared.py for data used by fragment writers
* Add BaseWriter class for common methods
* Add DomainWriter as an example
* Add imp.reloads for developing in modelbuilder
